Adding Visibility types to incoming federated posts
