+++ server/src/main/java/edu/sjsu/moth/server/controller/InboxController.java
@@ -13,6 +13,7 @@ import edu.sjsu.moth.server.db.AccountField;
 import edu.sjsu.moth.server.db.ExternalStatus;
 import edu.sjsu.moth.server.service.AccountService;
 import edu.sjsu.moth.server.service.ActorService;
+import edu.sjsu.moth.server.service.InboxService;
 import edu.sjsu.moth.server.service.StatusService;
 import edu.sjsu.moth.server.util.MothConfiguration;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -35,11 +36,9 @@ import java.util.List;
 
 @RestController
 public class InboxController {
-    @Autowired
-    StatusService statusService;
 
     @Autowired
-    ActorService actorService;
+    InboxService inboxService;
 
     @Autowired
     AccountService accountService;
@@ -51,71 +50,13 @@ public class InboxController {
         this.mappedLoad = mappedLoad;
     }
 
-    public static Mono<Account> convertToAccount(Actor actor) {
-        String serverName = "";
-        try {
-            serverName = new URL(actor.id).getHost();
-        } catch (MalformedURLException e) {
-            e.printStackTrace();
-        }
-
-        ArrayList<AccountField> accountFields = new ArrayList<>();
-        for (Attachment attachment : actor.attachment) {
-            AccountField accountField = new AccountField(attachment.name, attachment.value, null);
-            accountFields.add(accountField);
-        }
-
-        String iconLink;
-        if (actor.icon != null) {
-            iconLink = actor.icon.url;
-        } else {
-            iconLink = "";
-        }
-
-        String imageLink;
-        if (actor.image != null) {
-            imageLink = actor.image.url;
-        } else {
-            imageLink = "";
-        }
-
-        WebClient webClient =
-                WebClient.builder().defaultHeader(HttpHeaders.ACCEPT, "application/activity+json").build();
-        Mono<JsonNode> outboxResponse = webClient.get().uri(actor.outbox).retrieve().bodyToMono(JsonNode.class);
-        Mono<JsonNode> followersResponse = webClient.get().uri(actor.outbox).retrieve().bodyToMono(JsonNode.class);
-        Mono<JsonNode> followingResponse = webClient.get().uri(actor.outbox).retrieve().bodyToMono(JsonNode.class);
-        String finalServerName = serverName;
-        return outboxResponse.flatMap(jsonNodeOutbox -> {
-            int totalItems = jsonNodeOutbox.get("totalItems").asInt();
-            return followersResponse.flatMap(jsonNodeFollowers -> {
-                int totalItemFollowers = jsonNodeFollowers.get("totalItems").asInt();
-                return followingResponse.map(jsonNodeFollowing -> {
-                    int totalItemFollowing = jsonNodeFollowing.get("totalItems").asInt();
-                    //change avatar, avatar static, header, header static, last status to "" from iconLink and imageLink
-                    //change from String.valueOf(generateUniqueId()) to just their name
-                    //changed last status from null to actor.published
-                    return new Account(actor.preferredUsername, actor.preferredUsername,
-                                       actor.preferredUsername + "@" + finalServerName, actor.url, actor.name,
-                                       actor.summary, iconLink, iconLink, imageLink, imageLink,
-                                       actor.manuallyApprovesFollowers, accountFields, new CustomEmoji[0], false, false,
-                                       actor.discoverable, false, false, false, false, actor.published, actor.published,
-                                       totalItems, totalItemFollowers, totalItemFollowing);
-                });
-            });
-        });
-        //don't know about custom emojis, bot, and group
-        //noindex, moved, suspended, and limited are optional?
-        //icon, image, tag, attachment might be null
-        //not sure how to get last_status_at. outbox doesn't give a time, only the last status
-    }
-
     @PostMapping("/inbox")
     public Mono<ResponseEntity<Object>> inbox(@RequestBody JsonNode inboxNode) {
         String requestType = inboxNode.get("type").asText();
         if (requestType.equals("Delete")) {
             return Mono.empty();
         } else if (requestType.equals("Create")) {
-            return createHandler(inboxNode);
+            return inboxService.createHandler(inboxNode);
         } else if (requestType.equals("Update")) {
             System.out.println("I've seen UPDATE and it is going to be supported soon");
             return Mono.empty();
@@ -124,54 +65,19 @@ public class InboxController {
         }
     }
 
-    public Mono<ResponseEntity<Object>> createHandler(@RequestBody JsonNode node) {
-        JsonNode objNode = node.get("object");
-        String toLink = objNode.get("id").asText();
-
-        //putting in variables for now to make it easier to read
-        String id = toLink.substring(toLink.indexOf("/statuses/") + "/statuses/".length());
-        String createdAt = node.get("published").asText();
-        String inReplyTo = objNode.get("inReplyTo").asText();
-        Boolean sensitive = objNode.get("sensitive").asText().equals("true");
-        String language = objNode.get("contentMap").fields().next().getKey();
-        String content = objNode.get("content").asText();
-
-        //Making an actor and then converting to account
-        String accountLink = node.get("actor").asText();
-        return actorService.getActor(accountLink).switchIfEmpty(createActor(accountLink))
-                .flatMap(actor -> convertToAccount(actor)).flatMap(account -> {
-                    //not sure about spoiler text
-                    //haven't implemented media service yet, not sure about visibility
-                    //changed inreplyto to null
-                    ExternalStatus status =
-                            new ExternalStatus(null, createdAt, null, null, sensitive, "", "direct", language, null,
-                                               null, 0, 0, 0, false, false, false, false, content, null, null, account,
-                                               List.of(), List.of(), List.of(), List.of(), null, null, content,
-                                               node.get("published").asText());
-                    return statusService.saveExternal(status).map(ResponseEntity::ok);
-                });
-    }
-
-    public Mono<Actor> createActor(String accountLink) {
-        WebClient webClient =
-                WebClient.builder().defaultHeader(HttpHeaders.ACCEPT, "application/activity+json").build();
-        Mono<Actor> response = webClient.get().uri(accountLink).retrieve().bodyToMono(Actor.class);
-        return response.flatMap(actor -> actorService.save(actor));
-    }
-
     @PostMapping("/users/{id}/inbox")
     public Mono<String> usersInbox(@PathVariable String id, @RequestBody JsonNode inboxNode) {
         String requestType = inboxNode.get("type").asText();
         // follow or unfollow requests from other or same instances
         switch (requestType) {
             case "Follow" -> {
-                return accountService.followerHandler(id, inboxNode,false);
+                return accountService.followerHandler(id, inboxNode, false);
             }
-            case "Undo"->{
-                return accountService.followerHandler(id, inboxNode,true);
+            case "Undo" -> {
+                return accountService.followerHandler(id, inboxNode, true);
             }
             case "Accept" -> {
-                return accountService.acceptHandler(id,inboxNode);
+                return accountService.acceptHandler(id, inboxNode);
             }
             default -> {
                 return Mono.empty();
